Created new class TboClipLib
This class represent a TBO clip library and is created using a library
directory path string.

Inherits from GObject and every widget that work with a clip lib
should has an attribute with this object and ask for every data to
this.

All clip lib directory and file parsing should be done in this class.